Fix 404 Not Found in navigation-php-redirect demo
I don't know how this happened, but the original commit
in the jquery-mobile repo, correctly updated index.html when it
renamed redirect-target.html to redirect-target.php.

Ref https://github.com/jquery/jquery-mobile/pull/6534
Ref https://github.com/jquery/jquery-mobile/commit/d924303bf9a6

Yet, in the demos repo here, the copy somehow turned into a strange
hybrid where the file is renamed, but the reference is outdated.